+++ apps/meteor/client/lib/chats/uploads.ts
@@ -1,4 +1,4 @@
-import type { IRoom } from '@rocket.chat/core-typings';
+import type { IMessage, IRoom } from '@rocket.chat/core-typings';
 import { Emitter } from '@rocket.chat/emitter';
 import { Random } from '@rocket.chat/random';
 import fileSize from 'filesize';
@@ -6,6 +6,7 @@ import fileSize from 'filesize';
 import { getErrorMessage } from '../errorHandling';
 import type { UploadsAPI, EncryptedFileUploadContent } from './ChatAPI';
 import { isEncryptedUpload, type Upload } from './Upload';
+import { USER_ACTIVITIES, UserAction } from '../../../app/ui/client/lib/UserAction';
 import { fileUploadIsValidContentType } from '../../../app/utils/client';
 import { sdk } from '../../../app/utils/client/lib/SDKClient';
 import { i18n } from '../../../app/utils/lib/i18n';
@@ -14,10 +15,12 @@ import { settings } from '../settings';
 class UploadsStore extends Emitter<{ update: void; [x: `cancelling-${Upload['id']}`]: void }> implements UploadsAPI {
 	private rid: string;
 
-	constructor({ rid }: { rid: string }) {
-		super();
+	private tmid?: string;
 
+	constructor({ rid, tmid }: { rid: IRoom['_id']; tmid?: IMessage['_id'] }) {
+		super();
 		this.rid = rid;
+		this.tmid = tmid;
 	}
 
 	private uploads: readonly Upload[] = [];
@@ -54,6 +57,10 @@ class UploadsStore extends Emitter<{ update: void; [x: `cancelling-${Upload['id'
 
 	removeUpload = (id: Upload['id']): void => {
 		this.set(this.uploads.filter((upload) => upload.id !== id));
+
+		if (this.uploads.length === 0) {
+			UserAction.stop(this.rid, USER_ACTIVITIES.USER_UPLOADING, { tmid: this.tmid });
+		}
 	};
 
 	editUploadFileName = (uploadId: Upload['id'], fileName: Upload['file']['name']) => {
@@ -90,7 +97,10 @@ class UploadsStore extends Emitter<{ update: void; [x: `cancelling-${Upload['id'
 		}
 	};
 
-	clear = () => this.set([]);
+	clear = () => {
+		this.set([]);
+		UserAction.stop(this.rid, USER_ACTIVITIES.USER_UPLOADING, { tmid: this.tmid });
+	};
 
 	async send(file: File, encrypted?: EncryptedFileUploadContent): Promise<void> {
 		const maxFileSize = settings.peek('FileUpload_MaxFileSize');
@@ -177,7 +187,7 @@ class UploadsStore extends Emitter<{ update: void; [x: `cancelling-${Upload['id'
 
 				this.once(`cancelling-${id}`, () => {
 					xhr.abort();
-					this.set(this.uploads.filter((upload) => upload.id !== id));
+					this.removeUpload(id);
 					reject(new Error(i18n.t('FileUpload_Canceled')));
 				});
 			});
@@ -187,4 +197,5 @@ class UploadsStore extends Emitter<{ update: void; [x: `cancelling-${Upload['id'
 	}
 }
 
-export const createUploadsAPI = ({ rid }: { rid: IRoom['_id'] }): UploadsAPI => new UploadsStore({ rid });
+export const createUploadsAPI = ({ rid, tmid }: { rid: IRoom['_id']; tmid?: IMessage['_id'] }): UploadsAPI =>
+	new UploadsStore({ rid, tmid });