+++ apps/meteor/client/views/room/composer/messageBox/MessageBoxActionsToolbar/MessageBoxActionsToolbar.tsx
@@ -56,12 +56,11 @@ const MessageBoxActionsToolbar = ({
 	}
 
 	const room = useRoom();
-	const uploadsStore = tmid ? chatContext.threadUploads : chatContext.uploads;
 
 	const audioMessageAction = useAudioMessageAction(!canSend || typing || isRecording || isMicrophoneDenied, isMicrophoneDenied);
 	const videoMessageAction = useVideoMessageAction(!canSend || typing || isRecording);
-	const fileUploadAction = useFileUploadAction(!canSend || isRecording || isEditing, uploadsStore);
-	const webdavActions = useWebdavActions(!canSend || isRecording || isEditing, uploadsStore);
+	const fileUploadAction = useFileUploadAction(!canSend || isRecording || isEditing);
+	const webdavActions = useWebdavActions(!canSend || isRecording || isEditing);
 	const createDiscussionAction = useCreateDiscussionAction(room);
 	const shareLocationAction = useShareLocationAction(room, tmid);
 	const timestampAction = useTimestampAction(chatContext.composer);