+++ packages/ui-contexts/src/index.ts
@@ -106,7 +106,7 @@ export { useAccountsCustomFields } from './hooks/useAccountsCustomFields';
 export { useUserPresence } from './hooks/useUserPresence';
 export { useUnstoreLoginToken } from './hooks/useUnstoreLoginToken';
 export { useOnLogout } from './hooks/useOnLogout';
-export { useMediaDeviceMicrophonePermission, type requestDevice } from './hooks/useMediaDevicePermission';
+export { useMediaDeviceMicrophonePermission } from './hooks/useMediaDevicePermission';
 export { useWriteStream } from './hooks/useWriteStream';
 export { useUserCard } from './hooks/useUserCard';
 export type { SubscriptionWithRoom } from './types/SubscriptionWithRoom';