+++ packages/ddp-client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/ddp-client
 
+## 1.0.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@8.2.1
+  - @rocket.chat/rest-typings@8.2.1
+  - @rocket.chat/api-client@0.2.52
+  </details>
+
 ## 1.0.4
 
 ### Patch Changes